Fix is_patched check for Python 2 compatibility
In Python 2, after the patch is applied, the new init function is transparently wrapped in an unbound method, and so comparing `httplib.HTTPSConnection.__init__` directly with the `new_init` function will always fail. To fix this, compare `new_init` instead with the wrapped function.

Since both `patch()` and `unpatch()` call `is_patched()` before running, this bug makes it effectively impossible to toggle the patch in Python 2.